refactor: split session notification sender facade
