+++ src/hooks/session-notification-platform.ts
@@ -1,0 +1,22 @@
+import { platform } from "os"
+
+export type Platform = "darwin" | "linux" | "win32" | "unsupported"
+
+export function detectPlatform(): Platform {
+  const detected = platform()
+  if (detected === "darwin" || detected === "linux" || detected === "win32") return detected
+  return "unsupported"
+}
+
+export function getDefaultSoundPath(platform: Platform): string {
+  switch (platform) {
+    case "darwin":
+      return "/System/Library/Sounds/Glass.aiff"
+    case "linux":
+      return "/usr/share/sounds/freedesktop/stereo/complete.oga"
+    case "win32":
+      return "C:\\Windows\\Media\\notify.wav"
+    default:
+      return ""
+  }
+}
